Extract browser tag sidebar facade
